dose-v1/backend/requirements.txt
pydantic-settings=2.12.0
asyncpg=0.31.0
psycopg2=2.9.11
bcrypt=5.0.0
passlib=1.7.4
cffi=2.0.0 
cryptography=46.0.3 
ecdsa=0.19.1
pyasn1=0.6.1 
pycparser=2.23 
python-jose=3.5.0 
rsa=4.9.1 
six=1.17.0

dnspython=2.8.0 
email-validator=2.3.0 
idna=3.11

annotated-doc=0.0.4 
anyio=4.12.0
fastapi=0.127.0 
starlette=0.50.0   

